src-web/components/whisper_oai/requirements.txt
openai-whisper==20231117